Data-Migration-SQL-to-Json/requirements.txt
mysql-connector-python==8.2.0
pymongo==4.6.0
